feat: implement "load more songs" feature
Added a new function `loadMoreSongs` to the `App` component in `app.tsx` which is responsible for fetching the next set of songs and updating the `currentPage` state. This function is passed to the `SongList` component as a prop.

In the `SongList` component, used the `onEndReached` prop of the `FlatList` component to call the `loadMoreSongs` function when the user approaches the end of the list.

Adjusted the `fetchSongs` function to consider the `page` parameter when making the API request.